feat: update ACM certificate reference in main.tf and add import block in acm.tf
Updates ACM certificate reference and adds import block

Refactors the ACM certificate reference in the domain association settings to utilize the module output instead of a direct data reference.

Adds an import block in the ACM module for better management of the ACM certificate.

This change improves modularity and maintainability of the Terraform configuration.